Add files via upload
Created The Data Driven Dynamic Project on Blinkit Sales Dataset. Have tried to make it user friendly with Slicers and their Effects in the 
Existing Report Visuals. 
Along with them, have given solutions to a lot of Business Problems that can be given from the Data Set and represented in them in Visual.
These are some of those Business Problems:

Which Product Category have produced the most sales quantity?

What is the Growth Sales wise in the past and current year?

Which Payment Methods are mostly used by customers and how many customers
in numbers have used them to done the
payments so we could enhance our Payment Options Available?

Which Products are the Most Ordered Products
and the numbers of customers buying those products?

Does we have increased our total Customers Count or does they 
have decreased by this year?

What is the total sales by customers in the past and current year?

What is the Growth Customers wise in the past and current year?

Do we have a +ive or -ive on most of the orders we delivered?

On which part of Delivery Process Customers have given most feedbacks.
We shall look into that part of our delivery process?

What are the Ratings we have got as per our feedbacks?